% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAllFiles.R
\name{addAllFiles}
\alias{addAllFiles}
\title{Add a bunch of files to a data repository}
\usage{
addAllFiles(df)
}
\arguments{
\item{df}{A data frame of arguments to use in a call to \code{addFile}, one
column per argument. Column names are used to match arguments, not
positions. Extra columns ignored. Must have a column named \code{file}, and
that column may not contain duplicate files. Any \code{addFile} arguments
for which no columns are provided will use the normal \code{addFile}
defaults.}
}
\value{
A character vector of problems, one element (named) for each
  \code{file} in the input data frame. The value of the element is an empty
  string if the file was copied and verified to be copied without error, a
  missing value if the status of the copy is unknown, and a string describing
  the error if anything else went wrong.
}
\description{
This just wraps \code{\link{addFile}} so that a data frame of arguments can
be provided; \code{addFile} will be called iteratively with each one. The
vector of check results from each \code{addFile} call is compressed with
\code{\link{checkSummary}} into a single string, merging the result from all
attempted \code{addFile} operations into a character vector with one element
for each \code{file} in the provided data frame (named with the \code{file}
name) and a value string indicating success (empty) or describing failure.
}

