% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, world!}
\usage{
hello()
}
\value{
The printed string \code{'Hello, world!'}, invisibly.
}
\description{
This is an example function named \code{hello} which prints \code{'Hello,
world!'}.
}
\examples{
hello()
#=> Hello, world!

}

