% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkAllTo.R
\name{linkAllTo}
\alias{linkAllTo}
\title{Link a bunch of files to a data repository}
\usage{
linkAllTo(df)
}
\arguments{
\item{df}{A data frame of arguments to use in a call to \code{linkTo}, one
column per argument. Column names are used to match arguments, not
positions. Extra columns are ignored. Must have a column named \code{file},
and that column should not contain duplicate files. Any \code{linkTo}
arguments for which no columns are provided will use the normal
\code{linkTo} defaults.}
}
\value{
A character vector of problems, one element (named) for each
  \code{file} in the input data frame. The value of the element is an empty
  string if the file was linked to without error, a missing value if the
  status of the link is unknown, and a string describing the error if
  anything went wrong.
}
\description{
This just wraps \code{\link{linkTo}} so that a data frame of arguments can
be provided; \code{linkTo} will be called iteratively with each one. The
vector of check results from each \code{linkTo} call is compressed with
\code{\link{checkSummary}} into a single string, merging the result from all
attempted \code{linkTo} operations into a character vector with one element
for each \code{file} in the provided data frame (named with the \code{file}
name) and a value string indicating success (empty) or describing failure.
If a failure occurs with any file, a warning is given specifying the
number of files with problems.
}

