% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkTo.R
\name{linkTo}
\alias{linkTo}
\title{Create a symbolic link to a repository element.}
\usage{
linkTo(file, repo = getwd(), entry = NULL, asFile = basename(file),
  fileSize = NULL, checksum = NULL, checksumFunc = "tools::md5sum",
  onExists = "error", duplicateOk = TRUE)
}
\arguments{
\item{file}{The symbolic link to create. Assuming all optional elements are
specified, the target of this link in the repository will be
\code{<repo>\\<entry>\\<asFile>}. If a file object matching the link name
already exists, \code{duplicateOk} and \code{onExists} control what happens.}

\item{repo}{The base directory for the file to be linked to. By default will
use the current directory if not specified.}

\item{entry}{A subdirectory or subdirectory tree in the repo where the target
file is expected to be found. By default this is NULL, indicating the root
\code{repo} directory is used.}

\item{asFile}{The base name of the target file in the repository. By default
this is the base name of \code{file}. This target file must exist and be
a real file, not a link.}

\item{fileSize}{The expected size of the target file, in bytes. By default
this is \code{NULL}, meaning file size will not be checked. If a positive
value is supplied it will be checked against the size of the source file on
the file system. If it does not match then failure is reported and the
checksum is not checked.}

\item{checksum}{The expected checksum of the target file. By default this is
\code{NULL}, meaning no checksum is generated. If given it will be checked
agains the value provided by the \code{checksumFunc}.}

\item{checksumFunc}{The function or function name (as a string) that will be
used when calculating checksums. The calculated target file checksum will
be verified against any provided checksums before linking, and this may
also be used if \code{onExists} requires it. String function names may be
qualified with a \code{SomePackage::} prefix as the function named is
retrieved using \code{\link{getSomewhere}}. By default the function used is
'\code{'tools::md5sum'}. The specified function should have one parameter,
a file path. The returned checksum value should be an (atomic) vector type
but can not be \code{NULL} or a missing value. When \code{checksumFunc} is
called, \code{path} has already been verified and is known to exist on the
file system as a real file (not a directory or link).}

\item{onExists}{What to do if the link name already exists on the file
  system. May not be applied if the existing file is a duplicate of the
  target, as defined by \code{duplicateOk}.
\code{itemized}{
   \code{item}{\code{"error"} - Return failure result.}
   \code{item}{\code{"replaceFile"} - Delete existing file and replace with
   link. Generates a warning. If this is a link or a directory, this will
   fail with error.}
   \code{item}{\code{"replaceAny"} - Delete existing file system entry and
   replace with a link. Deletes recursively if the existing object is a
   directory. You probably don't want to do this. Generates a warning. }
   \code{item}{\code{"backupFile"} - Renames an existing file adding
   \code{.bak}, generating a warning. Multiple \code{.bak} may be added if
   needed to create unique filenames. Replaces the original with the link. If
   the original was a directory or a link, returns with error status. }
   \code{item}{\code{"backupAny"} - Renames the existing file system entry
   adding \code{.bak}, generating a warning. Multiple \code{.bak} may be
   added if needed to create unique filenames. Replaces the original with the
   link. This is done even if the original was a directory or a link.}
   \code{item}{\code{"skip"} - Leave as is. Generates a warning.}
}}

\item{duplicateOk}{By default \code{onExists} applies only if the file
exists and is not a duplicate. Set \code{FALSE} if \code{onExists} should
always be applied. Duplicate means is a real file and matches the target by
size and checksum, or is a link to the target. Links to an identical file
that is not in the repository is \emph{not} a duplicate. Real duplicate
files will be deleted and replaced with links. If the existing file is a
links to the target, nothing needs to be done.}
}
\value{
A named vector of validation results. Each element is named for a
  check performed, and will be the empty string if the check succeeded, a
  failure string if the check failed, and a missing value if the check was
  not performed (i.e. was not wanted or previous checks failed making further
  checking irrelevant.) Can be summarized with \code{\link{checkSummary}}.
  The elements are named \code{'checkParam_file', 'checkParam_repo',
  'checkParam_entry', 'checkParam_asFile', 'checkParam_checksumFunc',
  'checkParam_onExists', 'checkParam_duplicateOk', 'checkTarget',
  'checkExists', 'createLink', 'unexpectedError'}, in order.
}
\description{
Creates a symbolic link pointing to a target in the repository. The target
must exist, and if any validation information is included, the target must
match. What to do if the link to be created has the same name as an existing
file system object depends on \code{duplicateOk} and \code{onExists}. By
default failure will be returned unless the file system object is already a
link to the target, or if it is an exact copy of the duplicate. Successful
linking is reported as a vector of empty strings. Problems are reported with
one or more non-empty elements.
}

